+++ models/stg_shopify__order.sql
@@ -79,7 +79,7 @@ final as (
         billing_address_province_code,
         billing_address_zip,
         browser_ip,
-        total_shipping_price_set,
+        {{ shopify_source.json_to_string("total_shipping_price_set", ref('stg_shopify__order_tmp')) }} as total_shipping_price_set,
         shipping_address_address_1,
         shipping_address_address_2,
         shipping_address_city,